backend/requirements.txt
fastapi
uvicorn[standard]
chromadb
google-cloud-firestore
google-cloud-storage
python-dotenv
google-generativeai 